requirements.txt
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
